move price code out of token file
